Add login and logout logic

+++ backend/routes/customers.js
@@ -4,6 +4,10 @@ const pool = require('../db');
 
 // Endpoint pour récupérer les clients
 router.get('/', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   try {
     const result = await pool.query('SELECT * FROM customers');
     res.json(result.rows);
@@ -15,6 +19,10 @@ router.get('/', async (req, res) => {
 
 // Endpoint pour récupérer un client par ID
 router.get('/:id', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
   if (!id) {
     return res.status(400).send('ID parameter is required');
@@ -30,6 +38,10 @@ router.get('/:id', async (req, res) => {
 
 // Endpoint pour récupérer l'image d'un client par ID
 router.get('/customer_id/image', async (req,res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   var id;
   if (req.query.id) {
     id = req.query.id;
@@ -49,6 +61,10 @@ router.get('/customer_id/image', async (req,res) => {
 
 // Endpoint pour récupérer les historiques de paiement d'un client par ID
 router.get('/:id/payments_history', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
   if (!id) {
     return res.status(400).send('ID parameter is required');
@@ -64,6 +80,10 @@ router.get('/:id/payments_history', async (req, res) => {
 
 // Endpoint pour récupérer les vêtements d'un client par ID
 router.get('/:id/clothes', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
   if (!id) {
     return res.status(400).send('ID parameter is required');
@@ -79,6 +99,10 @@ router.get('/:id/clothes', async (req, res) => {
 
 // Endpoint pour supprimer les clients
 router.delete('/', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   try {
     await pool.query('DELETE FROM customers');
     res.json("All customers have been deleted");
@@ -90,6 +114,10 @@ router.delete('/', async (req, res) => {
 
 // Endpoint pour supprimer un client par ID
 router.delete('/:id', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
 
   if (!id) {
@@ -111,6 +139,10 @@ router.delete('/:id', async (req, res) => {
 
 // Endpoint pour supprimer l'image d'un client par ID
 router.delete('/:id/image', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
 
   if (!id) {
@@ -131,6 +163,10 @@ router.delete('/:id/image', async (req, res) => {
 
 // Endpoint pour supprimer les historiques de paiement d'un client par ID
 router.delete('/:id/payments_history', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
 
   try {
@@ -147,6 +183,10 @@ router.delete('/:id/payments_history', async (req, res) => {
 
 // Endpoint pour supprimer les vêtements d'un client par ID
 router.delete('/:id/clothes', async (req, res) => {
+  if (!global.ACCOUNT_TOKEN) {
+    console.log("The user is not connected")
+    return res.status(401).json({ error: 'Not connected' });
+  }
   const { id } = req.params;
 
   try {
